requirements.txt
streamlit
opencv-python
Pillow
ultralytics
streamlit-webrtc
